3 was done FC CardStarted for reflect WeaponsCard
CardStarted presence until user did not selecet weapons model
